Reuse uv code
